Scroll to after tab change
Perform actual scroll to group after current tab has been updated and is visible in the DOM. In this way it can get the actual offset of the group and scroll after changing the tab. Previously it changed tab but didn't scroll down to group as it return offset = 0; The getScrollPositionFor returns a timeout promise, with the group offset after it has been made visible.